+++ woodwork/datatable.py
@@ -764,11 +764,11 @@ class DataTable(object):
             data[col_name] = new_col.cat.codes
         return data
 
-    def get_mutual_information(self, num_bins=10, nrows=None):
+    def mutual_information(self, num_bins=10, nrows=None):
         """
-        Calculates mutual information between all pairs of columns in the DataTable
-        that support mutual information. Logical Types that support mutual information are
-        as follows:  Boolean, Categorical, CountryCode, Double, Integer, Ordinal, SubRegionCode, and ZIPCode
+        Calculates mutual information between all pairs of columns in the DataTable that
+        support mutual information. Logical Types that support mutual information are as
+        follows:  Boolean, Categorical, CountryCode, Double, Integer, Ordinal, SubRegionCode, and ZIPCode
 
         Args:
             num_bins (int): Determines number of bins to use for converting